pyproject.toml
[project]
name = "yandex-django"
version = "0.1.0"
dependencies = [
    "django",
    "djangorestframework>=3.16.1",
    "drf-yasg>=1.21.10",
    "markdown>=3.9",
    "pytest>=8.4.2",
    "pytest-django>=4.11.1",
    "ruff>=0.13.1",
]
requires-python = ">=3.12"
